pyproject.toml
[tool.poetry]
name = "med-ner-rag-trial"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
loguru = "^0.7.2"
asyncio = "^3.4.3"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
fastapi = {extras = ["standard"], version = "^0.115.4"}
haystack-ai = "^2.6.1"
ollama-haystack = "^1.1.0"
qdrant-haystack = "7.0.0"
fastembed-haystack = "^1.3.0"
onnxruntime = "1.19.2"
scikit-learn = "^1.5.2"
accelerate = "^1.1.1"
transformers = "^4.46.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-sugar = "^1.0.0"
pre-commit = "^4.0.1"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope="function"
filterwarnings = [
    "ignore::DeprecationWarning"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
